requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24
opencv-python-headless>=4.8.0
Pillow>=10.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.7
streamlit>=1.30
plotly>=5.18
tensorboard>=2.14
